requirements.txt
streamlit==1.35.0
pandas
gspread
oauth2client
matplotlib
requests
pillow
opencv-python-headless==4.8.0.74